backend/requirements.txt
fastapi
uvicorn
openai
elevenlabs
requests
moviepy
google-api-python-client
python-multipart
